join room

+++ src/messages/messages.service.ts
@@ -11,6 +11,8 @@ import { Board } from "src/board/board.model";
 import { Card } from "src/board/card.model";
 import { User } from "src/user/user.model";
 import { CreateSocketRoomDto } from "./dto/create-socket-room.dto";
+import { HelperService } from '../helper/helper.service';
+import { v4 as uuid } from "uuid";
 
 @Injectable()
 export class MessagesService {
@@ -22,7 +24,8 @@ export class MessagesService {
         @InjectModel(BoardCards) private boardCardRepository: typeof BoardCards,
         private jwtService: JwtService,
         private boardService: BoardService,
-        private userService: UserService
+        private userService: UserService,
+        private helperService: HelperService
     ) { }
 
 
@@ -44,6 +47,9 @@ export class MessagesService {
     async playCard(client: Socket, dto: PlayCardDto) {
         const user = await this.userRepository.findOne({ where: { socketId: client.id }, include: { all: true } })
         let check = await this.ruleChecker(dto.boardId, dto.cardId);
+        if (typeof check === 'number') {
+            return check;
+        }
         const cardDeleteFromUser = await this.userCardRepository.destroy({ where: { cardId: dto.cardId, userId: user.id } })
         const cardForBoard = await this.boardCardRepository.create({ cardId: dto.cardId, boardId: dto.boardId });
         return await cardForBoard.save();
@@ -60,6 +66,10 @@ export class MessagesService {
         if (cardsUsersArr.some((el) => el < cardId)) {
             throw new HttpException('GAME OVER!', HttpStatus.BAD_REQUEST);
         }
+        const nextLevelFunc = await this.helperService.nextLevel(boardId)
+        if (nextLevelFunc) {
+            return nextLevelFunc
+        }
         return true;
     }
 
@@ -69,21 +79,67 @@ export class MessagesService {
 
 
     async socketCreateRoom(client: Socket) {
-        if (!(client.handshake.query.token) || !(client.handshake.query.boardId)) {
+        if (!(client.handshake.query.token) || !(client.handshake.query.numberOfPlayers)) {
             throw new HttpException('Bad request params', HttpStatus.BAD_REQUEST);
         }
         if (Array.isArray(client.handshake.query.token)) {
             throw new HttpException('Bad request params', HttpStatus.BAD_REQUEST);
         }
+
         const userToken: string = client.handshake.query.token;
-        client.data.board = client.handshake.query.boardId;
-        await this.boardRepository.create()
         const user = await this.jwtService.verifyAsync(userToken, { secret: process.env.PRIVATE_KEY });
-        const boardId = client.handshake.query.boardId;
         const realUser = await this.userRepository.findOne({ where: { id: user.id }, include: { all: true } })
+
+        const roomMode = Number(client.handshake.query.numberOfPlayers);
+        const generatedPassword = uuid();
+        const board = await this.boardRepository.create({
+            boardPassword: generatedPassword, roomMode
+        });
+
+        await board.save();
+        const boardId = board.id;
+
+        client.data.board = boardId;
         realUser.socketId = client.id;
+
+        realUser.boardId = board.id;
+        await realUser.save();
+        board.createrUserId = realUser.id;
+
         await client.join(`${boardId}`);
-        return await realUser.save();
+        await realUser.save();
+
+        return { board, generatedPassword };
+    }
+
+    async joinRoom(client: Socket) {
+        if (!(client.handshake.query.token) || !(client.handshake.query.boardPassword)) {
+            throw new HttpException('Bad request params', HttpStatus.BAD_REQUEST);
+        }
+        if (Array.isArray(client.handshake.query.token) || Array.isArray(client.handshake.query.boardId)) {
+            throw new HttpException('Bad request params', HttpStatus.BAD_REQUEST);
+        }
+
+        const userToken: string = client.handshake.query.token;
+        const user = await this.jwtService.verifyAsync(userToken, { secret: process.env.PRIVATE_KEY });
+        const realUser = await this.userRepository.findOne({ where: { id: user.id }, include: { all: true } })
+
+        const boardPassword = client.handshake.query.boardPassword;
+        const board = await this.boardRepository.findOne({
+            where: { boardPassword: boardPassword }, include: { all: true }
+        });
+        if (!board) {
+            throw new HttpException('Wrong password', HttpStatus.BAD_REQUEST);
+        }
+        const boardId = board.id;
+
+        client.data.board = boardId;
+        realUser.socketId = client.id;
+        realUser.boardId = board.id;
+        await realUser.save();
+        await client.join(`${boardId}`);
+
+        return board;
     }
 
     async socketsLeave(socket: Socket, socketName: string) {
